Refatorando projeto.

+++ clinicamedica/src/main/java/br/com/projetos/clinicamedica/service/DoctorServiceImpl.java
@@ -7,6 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import br.com.projetos.clinicamedica.entity.Doctor;
+import br.com.projetos.clinicamedica.exception.ErrorRegisterNotFoundInDataBase;
 import br.com.projetos.clinicamedica.repository.DoctorRepository;
 
 @Service
@@ -24,25 +25,45 @@ public class DoctorServiceImpl implements DoctorService{
 	@Override
 	public Doctor findById(Long theId) {		
 		Optional<Doctor> result = doctorRepository.findById(theId);
-		
-		Doctor theDoctor = null;
+
 		if(result.isPresent()) {
-			theDoctor = result.get();
+			return result.get();
 		}else {
 			//Médico não encontrado
-			throw new RuntimeException("Noi possível encontrar o médico pelo id: " + theId);
+			throw new ErrorRegisterNotFoundInDataBase("Médico com Id " + theId + " não existe no banco de dados.");
 		}
-		return theDoctor;
 	}
 
 	@Override
 	public void save(Doctor theDoctor) {
+		// Caso o usuário envie um ID do frontend pelo JSON,
+		// esse resguardo seta ele como 0, para o sistema forçar a entender como um novo
+		// ao invés de fazer o update - Boas práticas!
+		theDoctor.setId(0L);
 		doctorRepository.save(theDoctor);		
 	}
 
 	@Override
-	public void deleteById(Long theId) {
-		doctorRepository.deleteById(theId);
+	public boolean deleteById(Long theId) {
+		Optional<Doctor> optionalDoctor = doctorRepository.findById(theId);
+
+		if(optionalDoctor.isPresent()){
+			doctorRepository.deleteById(theId);
+			return true;
+		}
+		throw new ErrorRegisterNotFoundInDataBase("Médico com Id " + theId + " não existe no banco de dados.");
+	}
+
+	@Override
+	public Doctor update(Doctor theDoctor) {
+		Optional<Doctor> optionalDoctor = doctorRepository.findById(theDoctor.getId());
+
+		if(optionalDoctor.isPresent()){
+			return doctorRepository.save(theDoctor);
+		}else{
+			throw new ErrorRegisterNotFoundInDataBase("Médico com Id " + theDoctor.getId() + " não existe no banco de dados.");
+		}
+		
 	}
     
 }
